Fix compiler warning, found on Mac
git-svn-id: https://subversion.gsi.de/go4/app/stream@3192 302e7798-3865-0410-85c9-d38765bbeff6

+++ framework/hadaq/HldFile.cxx
@@ -171,7 +171,7 @@ bool hadaq::HldFile::ReadBuffer(void* ptr, uint32_t* sz, bool onlyevent)
 
       // not possible to read event completely
       if ( readsz != (hdr->GetPaddedSize() - sizeof(hadaqs::HadTu))) {
-         fprintf(stderr, "HLD file reading problem: readsz %u not matches with padded size %u, abort reading\n", readsz, hdr->GetPaddedSize() - sizeof(hadaqs::HadTu));
+         fprintf(stderr, "HLD file reading problem: readsz %lu not matches with padded size %lu, abort reading\n", (long unsigned) readsz, (long unsigned) (hdr->GetPaddedSize() - sizeof(hadaqs::HadTu)));
          fEOF = true;
          return false;
       }
